requirements.txt
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
spotipy==2.25.1
uritemplate==4.1.1
urllib3==2.4.0
